+++ carp_communication_package/lib/communication.dart
@@ -7,6 +7,7 @@ import 'package:json_annotation/json_annotation.dart';
 import 'package:sms/sms.dart';
 import 'dart:async';
 import 'package:call_log/call_log.dart';
+import 'package:device_calendar/device_calendar.dart';
 import 'package:crypto/crypto.dart';
 import 'dart:convert';
 import 'package:carp_mobile_sensing/carp_mobile_sensing.dart';