+++ carp_mobile_sensing/lib/sampling_packages/apps/apps_datum.dart
@@ -9,7 +9,7 @@ part of apps;
 /// Holds a list of names of apps installed on the device.
 @JsonSerializable(fieldRename: FieldRename.snake, includeIfNull: false)
 class AppsDatum extends CARPDatum {
-  static const DataFormat CARP_DATA_FORMAT = DataFormat(NameSpace.CARP, DataType.APPS);
+  static const DataFormat CARP_DATA_FORMAT = DataFormat(NameSpace.CARP, AppsSamplingPackage.APPS);
   DataFormat get format => CARP_DATA_FORMAT;
 
   /// List of names on installed apps.
@@ -31,7 +31,7 @@ class AppsDatum extends CARPDatum {
 /// Holds a Map of names of apps and their corresponding usage in seconds.
 @JsonSerializable(fieldRename: FieldRename.snake, includeIfNull: false)
 class AppUsageDatum extends CARPDatum {
-  static const DataFormat CARP_DATA_FORMAT = DataFormat(NameSpace.CARP, DataType.APP_USAGE);
+  static const DataFormat CARP_DATA_FORMAT = DataFormat(NameSpace.CARP, AppsSamplingPackage.APP_USAGE);
   DataFormat get format => CARP_DATA_FORMAT;
 
   DateTime start, end;