+++ carp_mobile_sensing/lib/sampling_packages/apps/apps_package.dart
@@ -12,6 +12,11 @@ class AppsSamplingPackage implements SamplingPackage {
   Probe create(String type) {
     switch (type) {
       case APPS:
+        // there is an error in the device_apps plugin
+        // see https://github.com/g123k/flutter_plugin_device_apps/issues/12
+        // therefore the APPS probe is disabled right now.
+        // TODO - add the AppsProbe once the plugin is fixed.
+        //return null;
         return AppsProbe();
       case APP_USAGE:
         return AppUsageProbe();
@@ -28,13 +33,22 @@ class AppsSamplingPackage implements SamplingPackage {
     ..powerAware = true
     ..measures.addEntries([
       MapEntry(
-          DataType.APPS,
-          PeriodicMeasure(MeasureType(NameSpace.CARP, DataType.APPS),
-              name: 'Installed Apps', enabled: true, frequency: 24 * 60 * 60 * 1000)),
+          APPS,
+          PeriodicMeasure(
+            MeasureType(NameSpace.CARP, APPS),
+            // collect list of installed apps once pr. day
+            name: 'Installed Apps',
+            enabled: true,
+            frequency: 24 * 60 * 60 * 1000,
+          )),
       MapEntry(
-          DataType.APP_USAGE,
-          PeriodicMeasure(MeasureType(NameSpace.CARP, DataType.APP_USAGE),
-              name: 'Apps Usage', enabled: true, frequency: 60 * 60 * 1000, duration: 60 * 60 * 1000)),
+          APP_USAGE,
+          PeriodicMeasure(MeasureType(NameSpace.CARP, APP_USAGE),
+              // collect app usage every 10 min for the last 10 min
+              name: 'Apps Usage',
+              enabled: true,
+              frequency: 10 * 60 * 1000,
+              duration: 10 * 60 * 1000)),
     ]);
 
   SamplingSchema get light => common;
@@ -42,4 +56,26 @@ class AppsSamplingPackage implements SamplingPackage {
   SamplingSchema get minimum => common;
 
   SamplingSchema get normal => common;
+
+  SamplingSchema get debug => SamplingSchema()
+    ..type = SamplingSchemaType.DEBUG
+    ..name = 'Debugging app sampling schema'
+    ..powerAware = true
+    ..measures.addEntries([
+      MapEntry(
+          APPS,
+          PeriodicMeasure(MeasureType(NameSpace.CARP, APPS),
+              // collect list of installed apps once pr. day
+              name: 'Installed Apps',
+              enabled: true,
+              frequency: 60 * 1000)),
+      MapEntry(
+          APP_USAGE,
+          PeriodicMeasure(MeasureType(NameSpace.CARP, APP_USAGE),
+              // collect app usage every 10 min for the last 10 min
+              name: 'Apps Usage',
+              enabled: true,
+              frequency: 60 * 1000,
+              duration: 10 * 60 * 1000)),
+    ]);
 }