+++ carp_mobile_sensing/lib/sampling_packages/sensors/pedometer_probe.dart
@@ -7,27 +7,31 @@
 
 part of sensors;
 
-/// The pedometer probe listens to the hardware step counts.
+/// The pedometer probe listens to the hardware step count sensor.
 ///
 /// It samples step counts periodically, as specified by [frequency] in [PeriodicMeasure] and
 /// reports the step count in a [PedometerDatum] for the duration of the period.
 /// The probe reports the number of steps taken since the last sampling. For example, if [frequency]
 /// is 1 hour, then numbers of step are sampled and reported on an hourly basis.
-class PedometerProbe extends BufferingPeriodicStreamProbe {
+class PedometerProbe extends BufferingStreamProbe {
   int _steps = 0;
+  int _count = 0;
   DateTime _startTime = DateTime.now();
-  List<num> luxValues = new List();
 
+  // listen to the pedometer plugin - fires an event on every step
   Stream<dynamic> get bufferingStream => Pedometer().stepCountStream;
 
-  Future<Datum> getDatum() async => PedometerDatum(stepCount: _steps, startTime: _startTime, endTime: DateTime.now());
-
-  void onSamplingStart() {
-    _steps = 0;
+  Future<Datum> getDatum() async {
+    // check if we have passed midnight, then _steps has to be reset to zero.
+    if (_count < _steps) _steps = 0;
+    PedometerDatum pd = PedometerDatum(_count - _steps, _startTime, DateTime.now());
+    _steps = _count;
     _startTime = DateTime.now();
+    return pd;
   }
 
-  void onSamplingEnd() {}
-
-  void onSamplingData(count) => _steps += count;
+  // the pedometer plugin reports absolute number of steps taken since midnight.
+  void onSamplingData(count) {
+    _count = count;
+  }
 }