+++ carp_mobile_sensing_app/lib/src/sensing/sensing.dart
@@ -15,7 +15,10 @@ class Sensing {
     // register the sampling packages
     SamplingPackageRegistry.register(CommunicationSamplingPackage());
     SamplingPackageRegistry.register(ContextSamplingPackage());
-    SamplingPackageRegistry.register(AudioSamplingPackage());
+    // right now the audio plugin throws an exception
+    // see https://github.com/dooboolab/flutter_sound/issues/93
+    // TODO - enable once issue is solved.
+    //SamplingPackageRegistry.register(AudioSamplingPackage());
 
     // register data endpoints
     DataManagerRegistry.register(DataEndPointType.PRINT, new ConsoleDataManager());
@@ -30,14 +33,11 @@ class Sensing {
     // Get the study.
     study = await mock.getStudy(testStudyId);
 
-    print(study.toString());
-
     // Create a Study Controller that can manage this study, initialize it, and start it.
     controller = StudyController(study);
     //controller = StudyController(study, privacySchemaName: PrivacySchema.DEFAULT); // a controller w. privacy
     await controller.initialize();
     controller.start();
-    print("Sensing started ...");
 
     // listening on all data events from the study and print it (for debugging purpose).
     controller.events.forEach(print);
@@ -50,7 +50,6 @@ class Sensing {
   void stop() async {
     controller.stop();
     study = null;
-    print("Sensing stopped ...");
   }
 }
 
@@ -76,10 +75,16 @@ class StudyMock implements StudyManager {
         ..description =
             'This is a long description of a Study which can run forever and take up a lot of space and drain you battery and you have to agree to an informed consent which - by all standards - do not comply to any legal framework....'
         ..dataEndPoint = getDataEndpoint(DataEndPointType.PRINT)
-        ..dataFormat = NameSpace.OMH
-        ..addTask(Task()
-          ..measures =
-              SamplingSchema.common(namespace: NameSpace.CARP).getMeasureList([DataType.AUDIO, DataType.BLUETOOTH]));
+        //..dataFormat = NameSpace.OMH
+        ..addTask(Task()..measures = SamplingSchema.debug(namespace: NameSpace.CARP).measures.values.toList());
+
+//        ..addTask(Task()
+//          ..measures = SamplingSchema.common().getMeasureList(
+//              [DataType.TEXT_MESSAGE_LOG, DataType.LOCATION, DataType.ACTIVITY], namespace: NameSpace.CARP)
+//          ..addMeasure(CalendarMeasure(MeasureType(NameSpace.CARP, CommunicationSamplingPackage.CALENDAR),
+//              name: "Calendar", frequency: 11 * 1000, daysBack: 10, daysFuture: 10))
+//          ..addMeasure(
+//              WeatherMeasure(MeasureType(NameSpace.CARP, DataType.WEATHER), name: "Weather", frequency: 11 * 1000)));
 
 //    ..addTask(Task()..measures = SamplingSchema.common(namespace: NameSpace.CARP).measures.values.toList());
 
@@ -200,8 +205,8 @@ class StudyMock implements StudyManager {
   Task get appTask {
     if (_appTask == null)
       _appTask = Task("Application Task")
-        ..addMeasure(
-            PeriodicMeasure(MeasureType(NameSpace.CARP, DataType.APPS), name: "Installed Apps", frequency: 5 * 1000));
+        ..addMeasure(PeriodicMeasure(MeasureType(NameSpace.CARP, AppsSamplingPackage.APPS),
+            name: "Installed Apps", frequency: 5 * 1000));
 
     return _appTask;
   }
@@ -210,7 +215,7 @@ class StudyMock implements StudyManager {
   Task get appUsageTask {
     if (_appUsageTask == null) {
       _appUsageTask = Task("AppUsage Task")
-        ..addMeasure(PeriodicMeasure(MeasureType(NameSpace.CARP, DataType.APP_USAGE),
+        ..addMeasure(PeriodicMeasure(MeasureType(NameSpace.CARP, AppsSamplingPackage.APP_USAGE),
                 name: "App Foreground Usage Time", frequency: 10 * 1000, duration: 60 * 60 * 1000) // go back one hour
             );
     }
@@ -221,7 +226,7 @@ class StudyMock implements StudyManager {
   Task get audioTask {
     if (_audioTask == null) {
       _audioTask = Task("Audio Task")
-        ..addMeasure(AudioMeasure(MeasureType(NameSpace.CARP, DataType.AUDIO),
+        ..addMeasure(AudioMeasure(MeasureType(NameSpace.CARP, AudioSamplingPackage.AUDIO),
             name: "Ambient Audio",
             frequency: 10 * 1000, // once every 10 sec
             duration: 2 * 1000, // record 2 sec
@@ -239,12 +244,13 @@ class StudyMock implements StudyManager {
   Task get commTask {
     if (_commTask == null) {
       _commTask = Task("Communication Task")
-        ..addMeasure(PhoneLogMeasure(MeasureType(NameSpace.CARP, DataType.PHONE_LOG),
+        ..addMeasure(PhoneLogMeasure(MeasureType(NameSpace.CARP, CommunicationSamplingPackage.PHONE_LOG),
             name: "Phone Log", days: 10)) // 10 days of log
-        ..addMeasure(PeriodicMeasure(MeasureType(NameSpace.CARP, DataType.TEXT_MESSAGE_LOG),
+        ..addMeasure(PeriodicMeasure(MeasureType(NameSpace.CARP, CommunicationSamplingPackage.TEXT_MESSAGE_LOG),
             name: "SMS Message Log", frequency: 6 * 1000 // once every 10 sec
             ))
-        ..addMeasure(Measure(MeasureType(NameSpace.CARP, DataType.TEXT_MESSAGE), name: "SMS Text Messaging"));
+        ..addMeasure(Measure(MeasureType(NameSpace.CARP, CommunicationSamplingPackage.TEXT_MESSAGE),
+            name: "SMS Text Messaging"));
     }
     return _commTask;
   }
@@ -258,7 +264,8 @@ class StudyMock implements StudyManager {
   Task get connectivityTask {
     if (_connectivityTask == null) {
       _connectivityTask = Task("Connectivity Task")
-            ..addMeasure(Measure(MeasureType(NameSpace.CARP, DataType.CONNECTIVITY), name: "Connectivity"))
+            ..addMeasure(
+                Measure(MeasureType(NameSpace.CARP, ConnectivitySamplingPackage.CONNECTIVITY), name: "Connectivity"))
 //        ..addMeasure(PeriodicMeasure(MeasureType(NameSpace.CARP, DataType.BLUETOOTH),
 //            name: "Nearby Bluetooth Devices",
 //            frequency: 1 * 10 * 1000, // every minute
@@ -273,7 +280,8 @@ class StudyMock implements StudyManager {
   Task get contextTask {
     if (_contextTask == null) {
       _contextTask = Task("Context Task")
-        ..addMeasure(PeriodicMeasure(MeasureType(NameSpace.CARP, DataType.ACTIVITY), name: "Activity Recognition"));
+        ..addMeasure(PeriodicMeasure(MeasureType(NameSpace.CARP, ContextSamplingPackage.ACTIVITY),
+            name: "Activity Recognition"));
     }
     return _contextTask;
   }
@@ -282,7 +290,7 @@ class StudyMock implements StudyManager {
   Task get environmentTask {
     if (_environmentTask == null) {
       _environmentTask = Task("Environment Task")
-        ..addMeasure(WeatherMeasure(MeasureType(NameSpace.CARP, DataType.WEATHER),
+        ..addMeasure(WeatherMeasure(MeasureType(NameSpace.CARP, ContextSamplingPackage.WEATHER),
             name: "Local Weather",
             frequency: 1 * 30 * 1000, // once every minute
             apiKey: '12b6e28582eb9298577c734a31ba9f4f'));
@@ -297,10 +305,10 @@ class StudyMock implements StudyManager {
   Task get hardwareTask {
     if (_hardwareTask == null) {
       _hardwareTask = Task("Hardware Task")
-        ..addMeasure(PeriodicMeasure(MeasureType(NameSpace.CARP, DataType.MEMORY),
+        ..addMeasure(PeriodicMeasure(MeasureType(NameSpace.CARP, DeviceSamplingPackage.MEMORY),
             name: "Availabel Memory", frequency: 10 * 1000)) // 10 days of log
-        ..addMeasure(Measure(MeasureType(NameSpace.CARP, DataType.BATTERY), name: "Battery"))
-        ..addMeasure(Measure(MeasureType(NameSpace.CARP, DataType.SCREEN), name: "Screen Activity"));
+        ..addMeasure(Measure(MeasureType(NameSpace.CARP, DeviceSamplingPackage.BATTERY), name: "Battery"))
+        ..addMeasure(Measure(MeasureType(NameSpace.CARP, DeviceSamplingPackage.SCREEN), name: "Screen Activity"));
     }
     return _hardwareTask;
   }
@@ -309,7 +317,7 @@ class StudyMock implements StudyManager {
   Task get locationTask {
     if (_locationTask == null) {
       _locationTask = Task("Location Task")
-        ..addMeasure(Measure(MeasureType(NameSpace.CARP, DataType.LOCATION))..name = "Location");
+        ..addMeasure(Measure(MeasureType(NameSpace.CARP, ContextSamplingPackage.LOCATION))..name = "Location");
     }
     return _locationTask;
   }
@@ -318,7 +326,7 @@ class StudyMock implements StudyManager {
   Task get noiseTask {
     if (_noiseTask == null) {
       _noiseTask = Task("Noise Task")
-        ..addMeasure(NoiseMeasure(MeasureType(NameSpace.CARP, DataType.NOISE),
+        ..addMeasure(NoiseMeasure(MeasureType(NameSpace.CARP, AudioSamplingPackage.NOISE),
             name: "Ambient Noise",
             frequency: 30 * 1000, // How often to start a measure
             duration: 2 * 1000, // Window size
@@ -332,7 +340,7 @@ class StudyMock implements StudyManager {
   Task get pedometerTask {
     if (_pedometerTask == null) {
       _pedometerTask = Task("Pedometer Task")
-        ..addMeasure(PeriodicMeasure(MeasureType(NameSpace.CARP, DataType.PEDOMETER),
+        ..addMeasure(PeriodicMeasure(MeasureType(NameSpace.CARP, SensorSamplingPackage.PEDOMETER),
             name: "Pedometer (Step Count)", frequency: 1 * 20 * 1000 // how often to collect step count
             ));
     }
@@ -349,17 +357,17 @@ class StudyMock implements StudyManager {
     if (_sensorTask == null) {
       _sensorTask = Task("Sensor Task")
         ..addMeasure(PeriodicMeasure(
-          MeasureType(NameSpace.CARP, DataType.ACCELEROMETER),
+          MeasureType(NameSpace.CARP, SensorSamplingPackage.ACCELEROMETER),
           name: "Accelerometer",
           frequency: 8 * 1000, // How often to start a measure
           duration: 20, // Window size
         ))
-        ..addMeasure(PeriodicMeasure(MeasureType(NameSpace.CARP, DataType.GYROSCOPE),
+        ..addMeasure(PeriodicMeasure(MeasureType(NameSpace.CARP, SensorSamplingPackage.GYROSCOPE),
             name: "Gyroscope",
             frequency: 9 * 1000, // How often to start a measure
             duration: 100 // Window size
             ))
-        ..addMeasure(PeriodicMeasure(MeasureType(NameSpace.CARP, DataType.LIGHT),
+        ..addMeasure(PeriodicMeasure(MeasureType(NameSpace.CARP, SensorSamplingPackage.LIGHT),
                 name: "Ambient Light",
                 frequency: 11 * 1000, // How often to start a measure
                 duration: 700) // Window size