+++ carp_audio_package/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.5.0 BREAKING
+* **Breaking change.** This version has been migrated from the deprecated Android Support Library to *AndroidX*. 
+This should not result in any functional changes, but it requires any Android app using this plugin to also 
+[migrate](https://developer.android.com/jetpack/androidx/migrate) if they're using the original support library. 
+* Use [`carp_mobile_sensing`](https://pub.dartlang.org/packages/carp_mobile_sensing) version 0.5.0 
+* See Flutter [AndroidX compatibility](https://flutter.dev/docs/development/packages-and-plugins/androidx-compatibility)
+
 ## 0.3.5
 * upgrade to json_serializable v.2
 * fixed bug in audio recording probe (issue [#23](https://github.com/cph-cachet/carp.sensing-flutter/issues/23))